+++ boa_engine/src/builtins/array_buffer/tests.rs
@@ -11,5 +11,5 @@ fn ut_sunny_day_create_byte_data_block() {
 fn ut_rainy_day_create_byte_data_block() {
     let mut context = Context::default();
 
-    assert!(create_byte_data_block(usize::MAX, &mut context).is_err());
+    assert!(create_byte_data_block(u64::MAX, &mut context).is_err());
 }
